refactor: Renamed ActivePublicationPattern
Renamed ActivePublicationPattern to PublicationPattern
