--- src/components/SerialSections/ActivePublicationPattern/index.js
@@ -1,1 +1,0 @@
-export { default } from './ActivePublicationPattern';
