+++ src/components/SerialSections/index.js
@@ -1,5 +1,5 @@
 export { default as SerialInfo } from './SerialInfo';
 export { default as SerialPOLine } from './SerialPOLine';
-export { default as ActivePublicationPattern } from './ActivePublicationPattern';
+export { default as PublicationPattern } from './PublicationPattern';
 export { default as DeprecatedPublicationPatterns } from './DeprecatedPublicationPatterns';
 export { default as SerialPieceSets } from './SerialPieceSets';
